echo_find/requirements.txt
# core
streamlit>=1.20.0
torch>=1.13.0
transformers>=4.30.0
librosa>=0.10.0
soundfile>=0.11.0
numpy>=1.24.0
scipy>=1.10.0
pydub>=0.25.1
joblib>=1.3.0
tqdm>=4.65.0

# faiss (CPU). If you have a GPU and want the GPU version, install faiss-gpu per faiss docs.
faiss-cpu==1.7.4
# NOTE: faiss-cpu 1.7.4 is not available on some platforms/py versions. Using a more
# recent CPU wheel that is available on PyPI. If this fails on your system, use the
# conda-forge install shown in the README/notes below.
faiss-cpu==1.12.0
